Test change: portI2cRequiresSpecificWiring() GNSS message -> UBX-MON-VER. (#888)
The GNSS message which the portI2cRequiresSpecificWiring() test uses has always been UBX-MON-BATCH, which is a nice short fixed-length message that all GNSS modules support. Unfortunately, the ZED-F9P GNSS module doesn't, hence with this commit we switch to the rather larger, variable length, UBX-MON-VER message which all GNSS modules definitely support, and will continue to do so; fingers crossed.

Note: the test portSpiRequiresSpecificWiring(), which also uses the UBX-MON-BATCH message, is not moved to UBX-MON-VER in this PR because we don't currently have a ZED-F9P GNSS module connected to anything via SPI; it should be moved at some point though.