Test automation fix only: fix PyInvoke version in Docker image. (#891)
The latest version of "inv", or PyInvoke, which will be installed if the Ubuntu Docker image we use when testing is updated, is broken: it fails to find the "scripts" directory, a Python module where store some of our stuff.  Hence here we fix the version of PyInvoke to one that will work.