Test/automation changes only: Nordic nRF52/nRF53 related. (#892)
- Improved Segger JLink install/reliability:
After introducing test instance 26, an nRF52833DK board, a few months ago, we have been seeing frequent (~50%) failures to download to that board, the nrfjprog tool complaining that the Segger ARM JLink DLL is returning an unknown error (-1).  This seems to be fixed by updating to the latest version of both nrfjprog (10.21.0) and Segger JLink (v7.88b) (which also updates the FW on the debug chip of the board from September 2022 to November 2022); hence with this commit the Dockerfile used by the test system and the script files run by "inv" (when tests are run locally) are both updated to use these versions.

- Improved Segger JLink/NRF command-line tools installation: in addition, the same scripts will now download and install Segger Jlink on a Windows machine (before it would prompt the user to do so) and correctly detect whether Segger JLink is already installed on Linux (previously it would find a JLink installation even if it wasn't there).

Note that, as before, when you are running things via "inv", if there is _already_ an installation of the NRF command-line tools and Segger JLink then you will NOT be updated to a more recent tool version; in order to be updated to the new version automagically by "inv", please manually uninstall both the NRF command-line tools AND Segger JLink first.

- Improved NRF Connect installation on Windows: the pre-2.0 version of the NRF Connect tools used an entirely different installation mechanism for Windows, which disappeared after 2.0 was introduced, hence the installation process run by "inv" when used at the command-line on Windows is now updated to be the same as that on Linux, basically using "west" to do everything from Git.  With this NRF Connect 2.3.0 actually builds on Windows.

ALSO, with this commit, lib_common is no longer tested: it stopped building for Windows in the move to NRF Connect 2.3 and, since it is no longer in active use, it was not worth fixing.  The code remains though, in case it is of use in the future.