Test automation change, ESP-IDF: set target type. (#895)
Now that we are building for ESP32S3 and ESP32, the Espressif tool idf.py can end up complaining that the target is a different one, if it happens to build on the same machine for ESP32S3 after ESP32, and just fails the build.  To fix this, need to issue "idf.py set-target esp32s3" (or whatever) before the build.