CORRECTION TO PREVIOUS COMMIT (223ac1cba0a1fcda565885fc7d12beace31cb861) (#898)
The previous commit introduced a version of the C library function localtime_r().  However, since the implementation does not actually take into account conversion from UTC to local time (just from a time_t to a struct tm), it is _actually_ an implementation of gmtime_r(); calling it localtime_r() might turn out to be actively confusing in the future.

Hence the function is renamed: localtime_r() becomes gmtime_r() (before anyone notices :-)) and, with this, it no longer needs to be hooked into Zephyr (just Windows) as Zephyr has an implementation of gmtime_r().