BREAKING CHANGE if you have your own port: add uPortGetTimezoneOffsetSeconds(). (#897)
A function, uPortGetTimezoneOffsetSeconds(), is added to obtain the current time-zone offset in seconds.  A default implementation is provided in port/clib/u_port_timezone.c which returns zero: this will likely work in all cases except Windows and Linux.

This function is useful if a call is made to mktime() with a time value that is in UTC rather than the local time; mktime() assumes that its input is in local time and, since its output must be in UTC, it subtracts the time-zone offset (including daylight saving time as required).  If you are calling mktime() instead with a UTC time then you should add uPortGetTimezoneOffsetSeconds() to the result to get back to UTC.

IF YOU HAVE YOUR OWN PORT: a default implementation of this function, which will return zero, is included by ubxlib.cmake and ubxlib.mk.  If you do not use ubxlib.cmake or ubxlib.mk, or your system DOES have a concept of a time-zone (e.g. using newlib and calling settz()), you will need to provide an implementation of uPortGetTimezoneOffsetSeconds(); you may choose to copy the one provided in port/platform/zephyr/src/u_port.c, which should work for platforms that have a standard C library.