GNSS power-on change only: no reset at power-on when using native USB. (#905)
In the uGnssPwrOn() function the GNSS chip is reset to ensure a good clean state. However, if this is done with a GNSS chip that is being addressed over its native USB interface then the USB interface will also be reset, which is no good as it drops the connection.  Hence now, when the GNSS chip's native USB interface is in use, the reset is no longer applied.