Test change only: correct the implementation of uDeviceSerial test. (#909)
The test of uDeviceSerial wraps and re-uses the underlying uPortSerial.  However, when first coded, it simply cast the event callback function used by uPortSerial to that used by uDeviceSerial, which is a hack: uDeviceSerial takes a pointer to a structure as its "handle" parameter, whereas uPortSerial takes an int32_t handle, resulting in a few ugly warnings when compiling the test code.

This is now corrected; should help the up-coming 64-bit support.