Feature addition: GNSS AssistNow (MGA) and minimising TTFF. (#907)
This commit adds support for the GNSS AssistNow services (Online, Offline and Autonomous), plus other features required to minimise time to first fix (TTFF) for the case where the GNSS chip is connected directly to an MCU; where the GNSS chip is connected via an intermediate (e.g. cellular) module these features have always been supported (since the intermediate module performs all of the hard work): see the next commit for the ability to configure the intermediate module for AssistNow services more finely.

Note: the process of obtaining and then sending assistance data directly to the GNSS chip is a relatively heap-intensive one (probably around 10 kbytes minimum): you will need to be sure that you have enough heap available, only for a short while of course, to take advantage of this feature.  If this is a problem please let us know and we will look into optimisation measures.