Zephyr build change: location of Unity. (#916)
When building for Zephyr, Unity, which is required for the runner build and for testing, is now located as follows:

1.  if the environment variable UNITY_PATH is set, then Unity is taken from there, else...
2.  if Zephyr is being brought in with the nRF Connect SDK then Unity is taken from the nRF Connect SDK, else...
3.  the platform must include Unity (the single unity.c and the include path for the headers).

(1) is the new bit, which is now the default for the runner build and the test system.  This _should_ mean that native Zephyr can be used, i.e. outside of the nRF Connect SDK.