Test automation system change only: remove warnings about missing settings. (#917)
Before the most recent test system re-write, in the days when all test instances were hung off a single PC, it was important to know if any agent-specific setting was missing and hence a warning was printed out by the test system if that were the case. In the new multi-Pi architecture it is perfectly fine for an agent-specific setting to be missing if it is not a setting that particular agent happens to use. Hence the warning is now commented out: it is retained in case it should be needed for debugging in the future.