BREAKING CHANGE to DNS and Wifi captive portal APIs. (#915)
The optional callbacks to the uDns() and uWifiCaptivePortal() APIs are modified to be the other way up: the callback is a "keep going" callback, which should return true if the blocking API is to continue block, false if it is to exit (consistent with the other places in the code we use this pattern), rather than the current "exit" callback behaviour (which is the other way up).  Since this code is currently only on the development branch, is not yet published, and the `development` branch will be brought into `master` as a merge commit, this does not present an actual break in the `ubxlib` API.

A test is also now running for uWifiCaptivePortal() (on all test instances where a Wifi-capable module is attached). One largely cosmetic change is made to the core code, which is to pull the task stack/priority values out into the .h file so that the values are more public/visible and the user can override them at compile time without having to change the code, also set the task priority relatively, using the standard macro (U_CFG_OS_APP_TASK_PRIORITY + 1), as opposed to an absolute value.