AT Client and Wifi event task changes: increase stack sizes (#927)
When building ESP32 with PlatformIO, the pre-built ESP32 binary seems to need slightly larger task stacks; a stack canary can go off with the existing values for U_AT_CLIENT_CALLBACK_TASK_STACK_SIZE_BYTES/U_WIFI_MQTT_DATA_EVENT_STACK_SIZE of 2048/1526 bytes, hence they are increased to 2100/1600 bytes respectively.  The most peculiar thing is that we test exactly this build in the test system and there the stack canary does not go odd.  Odd.

Anyway, the sizes are hereby increased and we'll have to find a way to get these things to happen in test.

Also, the PlatformIO JSON file is updated so that the PlatformIO package manager is happy with it.