BIG CHANGE: 64-bit pointers and support for native Linux. (#928)
Support is added for running on native Linux (as opposed to through Zephyr/Posix, though that is still supported), including code and API updates (important bits highlighted below) to support Linux and 64-bit pointers.  In this initial commit all tests pass on Linux but with the following limitations:

- use of GPIOs is not yet implemented,
- no tests are yet being run on native Linux with a cellular/Wi-Fi/GNSS module attached in the automated test system.

See future commits for such additions.

API MODIFICATIONS FOR LINUX SUPPORT: a field, pPrefix, is added to the end of uDeviceSerial_t.  This is used in Linux only, for all other platforms it is IGNORED.

For Linux pPrefix may either be left at NULL or it may be populated with the prefix to the UART number, e.g. if pPrefix is "/dev/tty" and uart is 3 then the UART is "/dev/tty3".  If pPrefix is left at NULL U_PORT_UART_PREFIX (i.e. /dev/ttyUSB) will apply, if uart is negative then pPrefix alone will be applied.

API ADDITIONS FOR 64-BIT SUPPORT: your attention is drawn to three new functions in the AT Client:

- uAtClientAddExt() replaces uAtClientAdd(),
- uAtClientStreamGetExt() replaces uAtClientStreamGet(),
- uAtClientUrcHandlerHijackExt() replaces uAtClientUrcHandlerHijack().

The uAtClientXxxExt() versions of the functions take a pointer to a uAtClientStreamHandle_t structure as their parameter, rather than an int32_t handle; this allows properly-formed 64-bit pointers to be supported, not squished into 32-bit types.  All of the ubxlib code uses the uAtClientXxxExt() functions; support for the non-Ext() functions remains but may be deprecated and removed at some point in the future, and of course they will not work correctly on a 64-bit Linux machine.