Feature, GNSS: support for power-saving. (#931)
Support is added in the uGnssPwr API for all of the GNSS power saving features.  Note that these features are not supported by all GNSS devices: see details in the API itself.  Also added are a pair of functions, uGnssSetRetries()/uGnssGetRetries(), which allow communications with the GNSS device to be retried when there is no response, which could happen if the GNSS chip were in an inactive state due to power saving.

Plus additional documentation tweaks in the GNSS API.