Automation change only: corrected approach to GPIO access. (#934)
In the previous commit, access to GPIOs from the Docker image that we run inside during test automation was achieved by running the Docker image as the root user.  While this is fine, not a security risk (the test system is secured "at the gates" rather than when running the Docker image, which would be way too late), it _did_ mean that Git needed to have its safe directory check removed (since the repositories that the Docker image uses will have been created by the "ubxlib" user and are now being accessed by the "root" user) and removing the Git safe directory check _might_ not be desirable for others that choose to run our test system.

Hence with this commit access to GPIOs is done differently: the "ubxlib" user inside the Docker image now gets to be part of a new group, "gpio", which is given the same group number as the underlying Raspbian OS applies (997) and, hey presto, GPIOs are available inside the Docker image.