Cellular security fix and test change: make DTLS work. (#936)
The DTLS code was wrong: for a DTLS transfer the socket must first be connected and then uSockWrite()/uSockRead() must be used (our thanks to jraats for pointing this out). This was not spotted before because we did not have a DTLS test. This is now fixed, and a DTLS example is also added to make the usage pattern clear.

NOTE THAT THIS IS A BREAKING CHANGE FOR THE UBXLIB SOCKETS TEST ECHO SERVER: the echo server used for sockets testing has been updated to include DTLS and, in doing so, several things had to be changed, most importantly the keys used in the certificates for sockets testing are now elliptic curve since the Golang DTLS package we have used (pion) only supports elliptic curve ciphers for certificate-based authentication.  TO CONTINUE TO USE THE UBXLIB SOCKETS TEST ECHO SERVER WITH SECURITY ENABLED (PORT 5065, AND NOW PORT 5070) YOU MUST USE THIS CODE OR LATER.