BREAKING CHANGE to GNSS: streamed position to report non-fixes.
The streamed GNSS position APIs (uGnssPosGetStreamedStart()/uGnssPosGetStreamedStop() and uLocationGetContinuousStart()/uLocationGetStop() when run with GNSS) previously only called the user callback if an at-least-2D position fix had been achieved.

This is now changed to that the user-callback is called whenever a position message arrives from the GNSS device, irrespective of whether the message indicates that a position fix has been achieved.  In the no-fix or time-only-fix cases the user should check the value of the errorCode received by the callback: zero indicates a position fix, U_ERROR_COMMON_TIMEOUT indicates a message with no fix, which may be a time-only fix.  The parameters passed to the callback will be set to out of range values when they are not populated.

THE BREAKING PART: the user callback should always check the error code and the validity of the variables it is passed.