Test fix, httpClient: recover on network error. (#945)
Sometimes a module (e.g. cellular SARA-R4) can lose connectivity momentarily during the rather long HTTP test.  Such a loss of connectivity loses the HTTP context inside the module, leading to failures for the rest of the test, no matter how many retries are attempted.

To mitigate against this, the network status callback is now used to flag such an outage, allowing the test to set-up the HTTP context within the module once more before retrying.

Also, the HTTP echo server is update to print out the number of bytes it gets in a PUT/POST request, in order to monitor that [separate] error mode.