BREAKING CHANGE to uPortGpio API: add index for Linux. (#947)
The GPIO configuration structure gains an "index" element, which is used on Linux to allow the application to indicate the index of the GPIO chip a given pin is attached to e.g. for gpiochip0 on a BCM-based Raspberry Pi the index would be 0 (assuming U_PORT_GPIO_CHIP_NAME_BASE is "gpiochip", fixed at compile time).

The U_PORT_GPIO_CONFIG_DEFAULT and U_PORT_GPIO_SET_DEFAULT macros are updated to set a default index of -1, which will be interpreted as "use the default chip" (which is gpiochip0).

Provided you have always initialised any GPIO configuration structure to U_PORT_GPIO_CONFIG_DEFAULT, or if you are not using Linux, this constitutes no material change.

Note: this means that if you are setting an output that must begin high, and hence you call uPortGpioSet() _before_ uPortGpioConfig(), that pin must be on the default GPIO chip.