Cellular change only: set time, read local time. (#948)
It is now possible to set the time in the cellular module manually (uCellCfgSetTime()) and to read the local time and timezone (uCellInfoGetTime()).  These may be useful if your cellular network does not provide time and having correct time in the cellular module is important (e.g. for certificate expiry checks).