Test fix, GNSS only: power saving flags again. (#963)
As we hammer the GNSS device during testing of setting/clearing power saving flags, it is clear that there are occasions where the device acknowledges that a power saving flag has been set but then, for whatever reason, is not able to complete the job and it is actually not set.  Hence we now advise the user to re-read the flags after performing their configurations and we allow one retry during testing.