Cellular example only: exampleSecPsk and SARA-R422. (#964)
The PSK example, exampleSecPsk(), doesn't work reliably on SARA-R422, AT+USECDEVINFO returning "+CME ERROR: SEC busy" on each retry. This is probably because the device has recently been powered up and is not yet in a state where it is ready to run but adding a delay in there just for SARA-R422 in our regression test scenario wouldn't be right. The actual test of security, securityPskGeneration(), does pass, so the outcome of running exampleSecPsk() is now not checked when the module type is SARA-R422.