GNSS test debug only: print stats in MGA test. (#969)
We've been seeing some failures in the gnssMgaBasic() test where the GNSS device sends a NACK during restoration of the MGA database. The test is updated to print the communications stats as seen by the GNSS device when a failure occurs, to see if we are stressing it too much.

Also tidy a few lines of comment in u_lib_mga.c.