BREAKING CHANGE to assert hook-function type: puAssertFailed_t -> uAssertFailed_t. (#972)
The type that forms the assert hook was misnamed as puAssertFailed_t when it was originally created (the "p" is not necessary).  This commit corrects the named type to uAssertFailed_t; the original puAssertFailed_t type is retained but marked as deprecated and may be removed at some point in the future.