Test change only: fix rare bug in http_server.go.
Every so often, once a monthish, the HTTP test server fails with a segmentation fault "invalid memory address or nil pointer dereference".  Seems that this is in the background task which deletes empty directories; advice on the internet is that the "info" parameter which the anonymous function passed to the Go library function filepath.Walk() gets can, under some circumstances, be nil.  The HTTP server code now checks for this.  Hopefully no more segmentation faults.