Fix, GNSS power-saving and CFG-VAL: save to BBRAM also. (#976)
When GNSS on/off power saving mode is in use, the contents of RAM are lost on entry to power-saving mode, being restored from BBRAM on exit from power-saving mode.  Hence all CFG-VAL settings must by default be written to both RAM and BBRAM if that mode is to work.