Fix GCC truncation warning (GCC version 12.3.0 on Linux). (#981)
GCC version 12.3.0, as delivered with Ubuntu 22.04.2 LTS, has introduced a new and rather annoying warning where it will complain that a numeral printed into a buffer by snprintf() may be truncated even when you've used a format specifier to insist that it is truncated. For instance:

char buffer[4];
snprintf(buffer, sizeof(buffer), "%03d", (int) x);
...produces such a warning because buffer isn't big enough to store the string representation of a max-sized integer, even though %03d limits it to 3 digits; on Linux we have warnings as errors and so the build will fail.

The fix for this, avoiding pointlessly large buffers (in the above case the warning only goes away if buffer is made 12 bytes in size), is to grab the return value of snprintf() into a variable (so that GCC knows that you care about the printed length, which I guess must not take account of format specifiers (??)) and then ignore it. All a bit pointless but, unfortunately, necessary.