Fix to test automation only: tidy output. (#982)
The test report XML file can, sometimes, not be present (e.g. if the test instance is one where the build is carried out on the same machine as the tests and the build part failed, hence no tests were run).  The bit of Jenkinsfile that reports test results (using junit) now tolerates that and, hopefully, doesn't use old test data either.

Also some prints are added to clarify what has [not] failed when an error occurs.