Zephyr change only: compile if CONFIG_SERIAL, CONFIG_I2C or CONFIG_SPI are N. (#983)
On Zephyr ubxlib currently requires CONFIG_SERIAL, CONFIG_I2C and CONFIG_SPI to be defined for full functionality. However, a customer application may need only one of these, hence the ubxlib code now compiles for Zephyr if these are not included in the build. The error U_ERROR_COMMON_PLATFORM is returned when an attempt is made to open a port for which the relevant compilation flag in Zephyr has not been set.