MQTT BEHAVIOUR CHANGE for cellular: MQTT disconnect when there is no transport. (#987)
The code behind uCellMqttDisconnect() (and hence uMqttClientDisconnect()) would previously consider the return code in the URC from the cellular module, such that if disconnect returned "failed" (e.g. due to poor coverage), though the disconnect function would still return zero/success, a call to uCellMqttIsConnected() would return true ('cos that's what the broker would think). This does not match up with what happens internally within the cellular module, which is that, when asked to disconnect, even if it is unable to send the disconnect request to the broker, it would still perform a local MQTT disconnect.

The behaviour of the API is now modified to match the behaviour of the module: basically, provided that the module has responded with a URC to the uCellMqttDisconnect() command, whatever its contents, the connection is considered lost and uCellMqttIsConnected() will return false until a new connection is made.