GNSS MGA database test change only: workaround for QZSS issue. (#988)
There seems to be an issue with some GNSS devices (e.g. those of FW version 05.10): restoring an MGA database which includes a QZSS item in it can result in a NACK being returned by the GNSS device; the test code code is updated to ignore such an error and advice is added to u_gnss_mga.h about what to do if this occurs.