Test change only: enable cppcheck. (#992)
cppcheck was always enabled for code checker test instance 6.1, however it wasn't being installed in the Ubuntu Docker image and hence wasn't being run. It is now.

Note: with this latest Dockerfile build, instance 6.2, the STM32F4 instance of code checker, prints a message about not being able to find ldlogger.so in order to LD_PRELOAD it, I think because the path given to LD_PRELOAD is wrong (x86_64 v not). Though annoying, this is harmless, it does not affect the running of the check or the outcome; it can be ignored.