DEPRECATION warning: U_CFG_OS_CLIB_LEAKS (change to documentation only at the moment).
U_CFG_OS_CLIB_LEAKS has, for the last year or two, always been set to 0.  The ubxlib code still acts upon it; if U_CFG_OS_CLIB_LEAKS is set to 1 the code does not perform prints from dynamically created tasks, which _used_ to cause an _apparent_ (but not actual) memory leak on STM32F4.

However, since U_CFG_OS_CLIB_LEAKS is set to zero in all cases, the ubxlib code will be modified to act as if it were set to 0 and the macro will be removed at the start of 2024.

IF YOU HAVE YOUR OWN PORT and are setting U_CFG_OS_CLIB_LEAKS to 1 for some reason, be warned.