Test change only: HTTP test server file write. (#995)
We have been seeing very occasional cases of an HTTP PUT or POST apparently being truncated.  We had thought, up to now, that this might be because of a lack of flow control on the UART lines to a module causing data overruns.  However, we have now seen an occurrence of such a failure where the UART flow control lines are definitely connected, hence suspicion has fallen on the HTTP test server; here we use Golang and io.Copy() to write the received file to disk.  We were ignoring the error return value from this function since it is only writing to a local disk, how can that possibly go wrong, however there are hints on the internet that it can go wrong and hence we now try three times if io.Copy() returns a non-nil error (and print diagnostic information so that we get a log point when it does).