Fix subtle bug in uLocation, Wifi flavour: infinite loop. (#999)
This one took ages to find. When the static function wifiPosCallback() is called with Wifi position and we're going to go around again, if startAsync() happens not to be able to operate at that moment, the code should retry after a short delay with a guard count.

However, the bracketing of the while() loop condition was incorrect: rather than x being assigned to the return value of startAsync() it was being assigned to the value of the Boolean condition "startAsync() < 0". This resulted in x being 0 or 1 of course, rather than x being zero or a negative error return value, and it is only if x goes negative that the delay was added and the guard-counter incremented; hence the incorrect bracketing would lead to an infinite loop and potentially infinite allocations by startAsync(), depending on what was going on.

The code is now fixed.