Test fix, GNSS: avoid apparent [not actual] memory leak running GNSS alone. (#1004)
The test gnssMgaServer() uses cellular and a HTTPS connection, hence it needs to call uSecurityTlsCleanUp() to free the mutex that the security code creates in order to make the resource sums add up. This is a problem only if the test run is of just the GNSS tests (i.e. with a test filter "gnss"); otherwise uSecurityTlsCleanUp() is naturally called for other reasons.