Test change, ESP-IDF only: test with release 5.0.3. (#1005)
Release 5.0.1 of ESP-IDF had a problem where the use of the compiler "no return" directive would cause an error deep within the compiler boiler-plate headers (see https://github.com/espressif/esp-idf/issues/11339).  This has been fixed in ESP-IDF version 5.0.3 hence we (a) move to testing with ESP-IDF release 5.0.3 and (b) switch on warnings-as-errors for our "runner" test build.

Note: we tried moving to releases 5.1 and 5.1.1 but unfortunately a bug has been introduced in those releases which misinterprets the size of the flash inside the u-blox NINA-W15 module as 128 kbtyes instead of 8 Mbytes, meaning that the ESP32 fails to boot; this has been raised as an issue with ESP-IDF, see https://github.com/espressif/esp-idf/issues/12222.