GNSS change only: remove need to call uGnssPosGetStop() after uGnssPosGetStart(). (#1007)
Though unstated, it was necessary to call uGnssPosGetStop() before uGnssPosGetStart() could be called once more. This is now fixed, so uGnssPosGetStart() can be called at any time (the new request replacing the current one), the description of uGnssPosGetStart() has been updated to match and the test gnssPosPos() has also been updated to ensure that all of this remains true.