Fix for cellular MQTT and MQTT Client APIs: don't mutex-lock callbacks. (#1012)
The message indication and disconnect callbacks of the cellular MQTT API were coded such that the callbacks were thread-safe but in a way that also prevented any cellular APIs from being called from those callbacks; doing so would cause a perpetual mutex lock-up.

This is unnecessarily protective and so they are now coded such that you can do what you will, with no loss of thread-safety. Since we don't recommend calling uMqttClientMessageRead() from an MQTT Client message indication callback, for reasons explained above that function, the warning is updated to "IMPORTANT" so that people won't miss it (as now they won't end up in an irretrievable mutex-lock as a result).

Note that the Wifi case for the MQTT Client is unaffected by this change and always permitted such callbacks, so the two implementations now match.