MQTT Client fix only: permit clear of retained MQTT message. (#1013)
We missed a bit in the MQTT implementation, specifically:

"By default, every MQTT topic can have a retained message. The standard MQTT mechanism to clean up retained messages is sending a retained message with an empty payload to a topic. This will remove the retained message."

Setting "retain" for an MQTT message in uMqttClientPublish() is supported but that function also checks that the publish message is not empty, hence clearing a retained MQTT message was not possible. Calls to uMqttClientPublish() now permit an empty message if "retain" is set to true.

Our thanks to Michael Boeding for spotting this.