Test automation/runner change only: UBXLIB_FEATURES from DATABASE.md for CMake also. (#1019)
Previously, including UBXLIB_FEATURES as a definition in DATABASE.md would only have an effect on the UBXLIB_FEATURES compiled into a build for Make-based systems (i.e. STM32F4, nRF5SDK or the static-size builds) or if PlatformIO was in use.  It now also works for CMake-based systems (ESP-IDF, Linux, Windows and Zephyr) and has its own column in DATABASE.md, no longer hacked-in by being parsed out of the #defines.  To make room for this, the "duration" column is dropped from DATABASE.md; that field was introduced a few years ago for a version of the test system that we did not go ahead with, hence it can be safely removed.

This is necessary to support a 3rd party geodesic library in the up-coming GNSS geofencing feature.