Add user context to device handle
Global variables should be avoided, each callback registration should have a
user context argument to be passed to the callback when invoked.

There are callbacks that do not have a user context argument,
uGnssPosGetStreamedStart is an example.

This change introduces the ability to attach a user context to the device
handle, it will allow singleton callbacks such as uGnssPosGetStreamedStart to
be able to extract the context without refactoring the registration functions
and the callback signatures.

    void *context = "A";
    if (uDeviceOpen(&gDeviceCfg, &device_handle) == 0) {
        uDeviceSetUserContext(device_handle, context);
        assert(uDeviceGetUserContext(device_handle) == context);
    }

Signed-off-by: Alon Bar-Lev <alon.barlev@gmail.com>