Feature addition: Geofence. (#1027)
A common uGeofence API is added to the set of common APIs, enabling the creation of geofences, and uXxxGeofence APIs are added to each of the gnss, cell and wifi APIs to use those geofences.

This geofence implementation does NOT use the geofence facilities of a GNSS device, instead it provides a more flexible geofence mechanism (multiple fences consisting of multiple polygons, as well as multiple circles, inside or outside checks, pessimistic or optimistic checks, transit checks), which runs within ubxlib, on the MCU-side, and this means it can be applied to position obtained via GNSS, cellular or WiFi.

This API ONLY WORKS if the conditional compilation flag U_CFG_GEOFENCE is defined for your build: this is because it requires maths functions and floating-point types which would otherwise unnecessarily increase the linked size of the C run-time library. Zephyr users should note that the default Zephyr minimal C library will not work for this since it does not support maths functions: you will need newlib.

IMPORTANT: the default earth model is a spherical one, it does not take into account the true shape of the earth with WGS84 coordinates (used by all satellite systems), and hence is good for small geofences, e.g. fences with shapes that are no more than a kilometres in size. To use shapes larger than this with accuracy better than 0.5% (the worst-case error for spherical-earth calculations) an integration with https://github.com/geographiclib, as a sub-module, is offered; this will employ WGS84 coordinates and be good for shapes several thousand kilometres in size, though with a considerable increase in computational complexity; see the top of common/geofence/api/u_geofence_geodesic.h for guidance on how to do this, noting that it requires C++ support and exceptions support in your compiler.