STM32F4 platform change only: update newlib version warning.
The warning in heap_useNewlib.c about which version of newlib it has been verified with is updated from 3.0.0 to 3.3.0, which is the version of newlib that comes with ARM GCC version 10-2020-q4-major.  To be clear, we've not changed the compiler version we test with, this is just a tidy-up to remove the warning.