Advice on use of uDeviceClose(): powerOff set to true and a misbehaving device. (#1031)
As discussed at the tail end of this issue:

https://github.com/u-blox/ubxlib/pull/147

...the following scenario is possible:

- a device is opened and running, all is good,
- something goes wrong with the device, the ubxlib functions are returning error, it's gone titsup,
- as a recovery mechanism, the application calls uDeviceClose() with the powerOff flag set to true,
- this will return an error because the device is misbehaving, it cannot be powered off, and the resources associated with it WILL NOT HAVE BEEN FREE'D.

In this situation the correct recovery is to call uDeviceClose() again, this time with the powerOff flag set to false, to free the resources associated with the unresponsive device.  This logic always works because the only reason for an error to be returned by uDeviceClose(), assuming it has not been passed a wonky device handle, is if the device is misbehaving and has failed to respond to the power off request.

With this commit the documentation of the uDeviceClose() function is updated with this advice and all instances in the examples, tests, etc., where uDeviceClose() is called with true are updated to follow the advice.