Test fix, geofence: allow time for callback to complete. (#1032)
On fast platforms (e.g. Linux) it is possible for the URC callback to still not have exited when execution returns to the test task (because the callback has set a flag).  Give it a little time to do so in the test and re-jig the cellular geofence test to correctly count-up the number of geofence callbacks (there should be two, one for each fence), rather than using a simple flag.