Fix to GNSS/Cell location APIs: don't call geofence callback if no location is determined. (#1033)
GNSS and cellular do a test against a geofence in all cases, even if no location has been determined. The intention was that this would be signalled to the geofence callback by latitudeX1e9 and longitudeX1e9 being LLONG_MIN.  However, internally, what we have is both latitudeX1e7 and longitudex1e7 set to INT_MIN, and INT_MIN times 100 is not LLONG_MIN.  Since there is not a lot of point in testing an unknown position against a geofence, the GNSS and cellular code now doesn't bother doing the test (and therefore doesn't call the callback) if the latitude and longitude are unknown, which matches what the Wi-Fi location code was doing in any case.