Fix to uLocationGetXxx(): set defaults. (#1036)
Commit efdc592cb2aa840900e1c6591d00124f41eb8867 changed the behaviour of the uLocationGetXxx() APIs for GNSS to call their callbacks even if a navigation message arrived that did _not_ contain a fix, the presence of a fix being indicated by the errorCode parameter passed to the callback being zero.

However, the documentation was not updated to reflect this and the contents of pLocation passed to the callback were not set to sensible default values in the error case, which also meant that it was not possible to pick-up just a time fix (i.e. when there is no location fix, so the errorCode parameter would be non-zero, but there might still be a time fix, which could be checked-for by looking for a non-negative value in the timeUtc field of pLocation).

These issues are now fixed.