Cellular feature addition: allow AT command delay to be overridden. (#1039)
For each cellular module, a minimum time from the OK or ERROR being received for one AT command to the start of the next AT command is set, range usually between 20 and 100 milliseconds. It is now possible to read and override this value with uCellAtCommandDelayGet(), uCellAtCommandDelaySet(), should you wish to optimize the speed of execution of AT commands in your product.

VERY IMPORTANT: THIS SHOULD BE DONE WITH GREAT CARE; a delay that is too short may lead to AT interface failures, commands being confused with one another, etc., and these issues may occur in random places, at random times, dependent upon the AT command that was just executed and, potentially, external factors (network searching, GNSS chip interaction, etc.). You should only reduce the value from the default if you have a good set of regression tests for your product solution that you can run to verify that a shortened delay does not introduce problems for the AT commands that you end up using, in the situations that you use them.